task manager backend: usecase renamed to usecases
