+++ task-manager-backend/src/task/controllers/delete-task/delete-task.js
@@ -1,6 +1,6 @@
 const asyncHandler = require("../../../utils/async-handler");
 const ApiResponse = require("../../../utils/api-response");
-const { deleteTaskUsecase } = require("../../usecase");
+const { deleteTaskUsecase } = require("../../usecases");
 
 const { TASK_DELETE_SUCCESS } = require("../../../commons/constants");
