+++ task-manager-backend/src/user/controllers/get-all-user/get-all-user.js
@@ -1,6 +1,6 @@
 const asyncHandler = require("../../../utils/async-handler");
 const ApiResponse = require("../../../utils/api-response");
-const { getAllUserUsecase } = require("../../usecase");
+const { getAllUserUsecase } = require("../../usecases");
 
 const { PROFILE_FETCH_SUCCESS } = require("../../../commons/constants");
